refactor: move NotificationType to vo package
